+++ test/unit/sql/where.test.js
@@ -715,8 +715,8 @@ suite(Support.getTestDialectTeaser('SQL'), function() {
       });
     }
 
-    if (current.dialect.supports.JSON) {
-      suite('JSON', function () {
+    if (current.dialect.supports.JSONB) {
+      suite('JSONB', function () {
         test('sequelize.json("profile->>\'id\', sequelize.cast(2, \'text\')")', function () {
           expectsql(sql.whereItemQuery(undefined, this.sequelize.json("profile->>'id'", this.sequelize.cast('12346-78912', 'text'))), {
             postgres: "profile->>'id' = CAST('12346-78912' AS TEXT)"