+++ changelog.md
@@ -1,6 +1,7 @@
 # Future
 - [ADDED] include now supports string as an argument (on top of model/association), string will expand into an association matched literally from Model.associations
 - [FIXED] Accept dates as string while using `typeValidation` [#6453](https://github.com/sequelize/sequelize/issues/6453)
+- [ADDED] Sqlite JSON datatype support ([#6500](https://github.com/sequelize/sequelize/pull/6500))
 
 # 4.0.0-1
 - [CHANGED] Removed `modelManager` parameter from `Model.init()` [#6437](https://github.com/sequelize/sequelize/issues/6437)